# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TargetDesk API",
        "description": "API documentation for TargetDesk application",
        "contact": {
            "email": "admin@targetdesk.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List all categories",
                "description": "Get all categories with optional filtering by type, hierarchical structure",
                "operationId": "4e3d90d72b18f914836dc0476aa6c60c",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by category type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "secteur",
                                "taille",
                                "priorite",
                                "origine",
                                "personnalisee"
                            ]
                        }
                    },
                    {
                        "name": "hierarchical",
                        "in": "query",
                        "description": "Return hierarchical structure (true) or flat list (false)",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "include_children",
                        "in": "query",
                        "description": "Include children categories in response",
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "description": "Create a new category with optional parent for hierarchical structure",
                "operationId": "05f9feb5943fc644eb52f8195a6b310c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Industrie Automobile"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Secteur de l'industrie automobile"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#FF6B6B"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "secteur",
                                            "taille",
                                            "priorite",
                                            "origine",
                                            "personnalisee"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/v1/clients/{clientId}/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get client categories",
                "description": "Get all categories assigned to a specific client",
                "operationId": "a080b0503cb3e7a36990ae2bf1b9abc4",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client categories retrieved successfully"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Assign categories to client",
                "description": "Assign one or multiple categories to a client",
                "operationId": "603efd942aa8460e3fe6280837c030e7",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_ids"
                                ],
                                "properties": {
                                    "category_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            3,
                                            5
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categories assigned successfully"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/clients/{clientId}/categories/{categoryId}": {
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Remove category from client",
                "description": "Remove a specific category assignment from a client",
                "operationId": "516cad753ebed9c453bac1f9086537de",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category removed from client successfully"
                    },
                    "404": {
                        "description": "Client or category not found, or not assigned"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}